Move provider commit test
